Solved problem with query search on DB
When creating a query search with NoSQL we need to choose the container to use based on the table choosen in the query.

I also have a problem with docker as I cannot upload blobs, for some reason my host-name is docker-desktop and I don't know how to change it.

delete method on Short still has a problem that might be due to the blob not existing yet, but essencialy when we get a token to compare on the deletes the hashes are diferent